profiling
Macro finish_frame

macro_rules! finish_frame {
    () => { ... };
}

Finishes the frame. This isn’t strictly necessary for some kinds of applications but a pretty normal thing to track in games.