Crate profiling

Macros

finish_frame

Finishes the frame. This isn’t strictly necessary for some kinds of applications but a pretty normal thing to track in games.

function_scope

Opens a scope automatically named after the current function.

register_thread

Registers a thread with the profiler API(s). This is usually setting a name for the thread. Two variants:

scope

Opens a scope. Two variants:

Attribute Macros

all_functions

Proc macro for creating a scope around each function under struct impl block

function

Proc macro for creating a scope around the function, using the name of the function for the scope’s name

skip

Proc macro to skip the auto_impl for the function