Crate profiling

Modules

empty_impl 🔒

Macros

finish_frame

Finishes the frame. This isn’t strictly necessary for some kinds of applications but a pretty normal thing to track in games.

function_scope

Opens a scope automatically named after the current function.

register_thread

Registers a thread with the profiler API(s). This is usually setting a name for the thread. Two variants:

scope

Opens a scope. Two variants: